Task_2/requirements.txt
tensorflow
matplotlib
numpy
Streamlit
pillow